Trung anh

+++ TTAP-TShirt/src/main/java/com/project/ttaptshirt/repository/KhuyenMaiRepo.java
@@ -3,5 +3,5 @@ package com.project.ttaptshirt.repository;
 import com.project.ttaptshirt.entity.KhuyenMai;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface KhuyenMaiRepo extends JpaRepository<KhuyenMai, Integer> {
+public interface KhuyenMaiRepo extends JpaRepository<KhuyenMai, Long> {
 }
